+++ base/exports.jl
@@ -836,6 +836,7 @@ export
     close,
     closewrite,
     countlines,
+    DirEntry,
     eachline,
     readeach,
     eof,