+++ stdlib/REPL/src/REPLCompletions.jl
@@ -8,7 +8,6 @@ using Core: Const
 const CC = Core.Compiler
 using Base.Meta
 using Base: propertynames, something, IdSet
-using Base.Filesystem: _readdirx
 
 abstract type Completion end
 
@@ -335,7 +334,7 @@ function cache_PATH()
         end
 
         path_entries = try
-            _readdirx(pathdir)
+            readdir(DirEntry, pathdir)
         catch e
             # Bash allows dirs in PATH that can't be read, so we should as well.
             if isa(e, Base.IOError) || isa(e, Base.ArgumentError)
@@ -398,9 +397,9 @@ function complete_path(path::AbstractString;
     end
     entries = try
         if isempty(dir)
-            _readdirx()
+            readdir(DirEntry)
         elseif isdir(dir)
-            _readdirx(dir)
+            readdir(DirEntry, dir)
         else
             return Completion[], dir, false
         end
@@ -1435,7 +1434,7 @@ function completions(string::String, pos::Int, context_module::Module=Main, shif
                     complete_loading_candidates!(suggestions, s, dir)
                 end
                 isdir(dir) || continue
-                for entry in _readdirx(dir)
+                for entry in readdir(DirEntry, dir)
                     pname = entry.name
                     if pname[1] != '.' && pname != "METADATA" &&
                         pname != "REQUIRE" && startswith(pname, s)